Update MainActivity.java
Changed the code to be able to do more than one operation, ex. 5+3+8 = 16 now. You can do the same operations as many times as needed. Does not work with a combination of other operations. ex 5*2/5 does not work. Might be because of something to do with val_one not being updated when switching operations. ex 5+2 /3/3 will give you 1, which could mean that it only remembers 3/3.